+++ rio/test/RIO/Prelude/SimpleSpec.hs
@@ -1,0 +1,13 @@
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE ViewPatterns #-}
+module RIO.Prelude.SimpleSpec (spec) where
+
+import RIO
+import RIO.Process
+import Test.Hspec
+
+spec :: Spec
+spec = do
+  it "logging works" $ asIO $ runSimpleApp $ logDebug "logging allowed"
+  it "process calling works" $ asIO $ runSimpleApp $ proc "echo" ["hello"] runProcess_