+++ rio/ChangeLog.md
@@ -2,8 +2,14 @@
 
 ## 0.1.3.0
 
+* Add `newLogFunc` function to create `LogFunc` records outside of a callback scope
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
 * Add `foldMapM`
+* Add `headMaybe`, `lastMaybe`, `tailMaybe`, `initMaybe`, `maximumMaybe`, `minimumMaybe`,
+  `maximumByMaybe`, `minimumByMaybe` functions to `RIO.List` module (issue #82)
+* Move non partial functions `scanr1` and `scanl1` from `RIO.List.Partial` to `RIO.List` (issue #82)
+* Add `SimpleApp` and `runSimpleApp`
+* Add `asIO`
 
 ## 0.1.2.0